requirements.txt
python-telegram-bot==21.0.1
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3